separate the categories and runners a bit.
there's now a Categories tab instead of an All one. and it means it now
views results from the sal only when in that tab. this is a big change
in terms of behavior and i'm not sure it's what we want, but it just
might be..so we'll have to see